+++ lib/td_dq/search/mappings.ex
@@ -4,6 +4,7 @@ defmodule TdDq.Search.Mappings do
   """
   alias TdCache.TemplateCache
 
+  @raw %{raw: %{type: "keyword"}}
   @raw_sort %{raw: %{type: "keyword"}, sort: %{type: "keyword", normalizer: "sortable"}}
 
   def get_mappings do
@@ -17,7 +18,7 @@ defmodule TdDq.Search.Mappings do
         type: "nested",
         properties: %{
           id: %{type: "long"},
-          name: %{type: "text", fields: %{raw: %{type: "keyword"}}}
+          name: %{type: "text", fields: @raw}
         }
       },
       rule_type_id: %{type: "long"},
@@ -29,8 +30,8 @@ defmodule TdDq.Search.Mappings do
       updated_by: %{
         properties: %{
           id: %{type: "long"},
-          user_name: %{type: "text", fields: %{raw: %{type: "keyword"}}},
-          full_name: %{type: "text", fields: %{raw: %{type: "keyword"}}}
+          user_name: %{type: "text", fields: @raw},
+          full_name: %{type: "text", fields: @raw}
         }
       },
       current_business_concept_version: %{
@@ -55,14 +56,14 @@ defmodule TdDq.Search.Mappings do
       },
       type_params: %{
         properties: %{
-          name: %{fields: %{raw: %{type: "keyword"}}, type: "text"}
+          name: %{fields: @raw, type: "text"}
         }
       },
       execution_result_info: %{
         properties: %{
           result_avg: %{type: "long"},
           last_execution_at: %{type: "date", format: "strict_date_optional_time||epoch_millis"},
-          result_text: %{type: "text", fields: %{raw: %{type: "keyword"}}}
+          result_text: %{type: "text", fields: @raw}
         }
       },
       _confidential: %{type: "boolean"},
@@ -122,7 +123,7 @@ defmodule TdDq.Search.Mappings do
   end
 
   defp mapping_type(values) when is_map(values) do
-    %{type: "text", fields: %{raw: %{type: "keyword"}}}
+    %{type: "text", fields: @raw}
   end
 
   defp mapping_type(_default), do: %{type: "text"}